Move grid-specific helper into generateGrid.ts
